removed old UTTestController.cs that was creating deprecation warnings and replaced it with a more up to date CustomUISample behaviour
